Moves greeting class into the PlayAgainInput class
